# api/docs/swagger.yaml
basePath: /local-eats
definitions:
...
  kitchen.CreateRequest:
    properties:
      address:
        type: string
      cuisine_type:
        type: string
      description:
        type: string
      name:
        type: string
      owner_id:
        type: string
      phone_number:
        type: string
    type: object
  kitchen.CreateResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      cuisine_type:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      phone_number:
        type: string
      rating:
        type: number
    type: object
  kitchen.Info:
    properties:
      address:
        type: string
      created_at:
        type: string
      cuisine_type:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      phone_number:
        type: string
      rating:
        type: number
      total_orders:
        type: integer
      updated_at:
        type: string
    type: object
  kitchen.KitchenDetails:
    properties:
      cuisine_type:
        type: string
      id:
        type: string
      name:
        type: string
      rating:
        type: number
      total_orders:
        type: integer
    type: object
  kitchen.Kitchens:
    properties:
      kitchens:
        items:
          $ref: '#/definitions/kitchen.KitchenDetails'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  kitchen.NewDataNoID:
    properties:
      description:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  kitchen.UpdatedData:
    properties:
      address:
        type: string
      cuisine_type:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      phone_number:
        type: string
      rating:
        type: number
      updated_at:
        type: string
    type: object
  order.Item:
    properties:
      dish_id:
        type: string
      quantity:
        type: integer
    type: object
  order.ItemDetails:
    properties:
      dish_id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
  order.NewOrder:
    properties:
      delivery_address:
        type: string
      delivery_time:
        type: string
      items:
        items:
          $ref: '#/definitions/order.Item'
        type: array
      kitchen_id:
        type: string
      user_id:
        type: string
    type: object
  order.NewOrderResp:
    properties:
      created_at:
        type: string
      delivery_address:
        type: string
      delivery_time:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/order.Item'
        type: array
      kitchen_id:
        type: string
      status:
        type: string
      total_amount:
        type: number
      user_id:
        type: string
    type: object
  order.OrderCustomer:
    properties:
      delivery_time:
        type: string
      id:
        type: string
      kitchen_name:
        type: string
      status:
        type: string
      total_amount:
        type: number
    type: object
  order.OrderInfo:
    properties:
      created_at:
        type: string
      delivery_address:
        type: string
      delivery_time:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/order.ItemDetails'
        type: array
      kitchen_id:
        type: string
      kitchen_name:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  order.OrderKitchen:
    properties:
      delivery_time:
        type: string
      id:
        type: string
      status:
        type: string
      total_amount:
        type: number
      user_name:
        type: string
    type: object
  order.OrdersCustomer:
    properties:
      limit:
        type: integer
      orders:
        items:
          $ref: '#/definitions/order.OrderCustomer'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  order.OrdersKitchen:
    properties:
      limit:
        type: integer
      orders:
        items:
          $ref: '#/definitions/order.OrderKitchen'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  order.StatusNoID:
    properties:
      status:
        type: string
    type: object
  order.UpdatedOrder:
    properties:
      id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway of Local Eats
  title: Local Eats
  version: "1.0"
paths:
...
  /kitchens:
    get:
      description: Fetches all kitchens from database
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kitchen.Kitchens'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Fetches all kitchens
      tags:
      - kitchen
    post:
      description: Inserts a new kitchen into database
      parameters:
      - description: Kitchen info
        in: body
        name: kitchen
        required: true
        schema:
          $ref: '#/definitions/kitchen.CreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kitchen.CreateResponse'
        "400":
          description: Invalid kitchen data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a kitchen
      tags:
      - kitchen
  /kitchens/{id}:
    delete:
      description: Deletes kitchen from database
      parameters:
      - description: Kitchen ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid kitchen ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes a kitchen
      tags:
      - kitchen
    get:
      description: Retrieves kitchen info from database
      parameters:
      - description: Kitchen ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kitchen.Info'
        "400":
          description: Invalid kitchen ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets a kitchen
      tags:
      - kitchen
    put:
      description: Updates kitchen info in database
      parameters:
      - description: Kitchen ID
        in: path
        name: id
        required: true
        type: string
      - description: Kitchen info
        in: body
        name: kitchen
        required: true
        schema:
          $ref: '#/definitions/kitchen.NewDataNoID'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kitchen.UpdatedData'
        "400":
          description: Invalid kitchen ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates a kitchen
      tags:
      - kitchen
...
  /kitchens/{id}/orders:
    get:
      description: Gets orders from database
      parameters:
      - description: Filter
        in: query
        name: filter
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrdersKitchen'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets orders for kitchen
      tags:
      - order
...
  /kitchens/search:
    get:
      description: Searches kitchens from database
      parameters:
      - description: Search query
        in: query
        name: query
        type: string
      - description: Cuisine type
        in: query
        name: cuisine_type
        type: string
      - description: Rating
        in: query
        name: rating
        type: number
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kitchen.Kitchens'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Searches kitchens
      tags:
      - kitchen
  /orders:
    get:
      description: Gets orders from database
      parameters:
      - description: Pagination
        in: query
        name: pagination
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrdersCustomer'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets orders for customer
      tags:
      - order
    post:
      description: Inserts a new order into database
      parameters:
      - description: Order info
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.NewOrder'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.NewOrderResp'
        "400":
          description: Invalid order data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates an order
      tags:
      - order
  /orders/{id}:
    get:
      description: Gets order from database
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrderInfo'
        "400":
          description: Invalid order ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets an order
      tags:
      - order
  /orders/{id}/status:
    put:
      description: Updates order status in database
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/order.StatusNoID'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.UpdatedOrder'
        "400":
          description: Invalid order ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates an order
      tags:
      - order
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"